GUI & better parser
I improved the parser by adding okay error messages and making it
whitespace sensitive. This means that it now works properly with both
line breaks and semi colons.

I also wrote a PNG output system. It's not very sophisticated, but one
of the nice side effects is that it allowed me to write a version of
the REPL which displays what you construct as you construct it.

The REPL needs some love right now (how to deal with givens and
includes?)