Primitive Destructuring & Non-Point Givens
Primitves may now be destructure with the names `circle`, `line`,
`segment`, and `ray`.

Now when the user clicks a point is produced on the canvas, *but is not
registered* as a given for the construction.

To declare the givens for a construction the user says

```
given point A, line l, ...
```

which one might see them doing at the same time that they declare
returns.

The `:write` command verifies that the returns only depend on the
declared givens.