Extracted Builtins, cleaned interpreter, bugfixes
The logic for builtins has been extracted into a new object.

The interpreter is much cleaner as the result of this and other changes.

Error reporting is improved, and I began to lay the foundations for
internal/external errors. Right now the plan is to throw external
errors and return internal ones.

Fixed a bug with line equivalence. Fixed a bug with intersections being
needlessly wrapped in product values.